Refactor Jucewright browser runner
