+++ scripts/jucewright_osci_browser/__init__.py
@@ -1,0 +1,5 @@
+"""Reusable Jucewright automation helpers for osci-render E2E browsing."""
+
+from .scenario import OsciRenderBrowserRun
+
+__all__ = ["OsciRenderBrowserRun"]
