+++ scripts/jucewright_osci_browser/errors.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+class StepError(RuntimeError):
+    pass
